Updated to latest version of X-Code and Separated JSON Error Objects into separate file for more modular code organization
